Update for examples, remove fire as another already exists
